+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # 0.0.4
 - Preserve entry list scroll position in Manager when adding, deleting, or reordering entries
+- Save Manager scroll position per user and timeline, automatically restored on next view
 - Save viewer scroll position per user and timeline, automatically restored on next view
 - Fix Deprecated warning
+- Small CSS fix for renderJournalDirectory